only expose thickness and fill as keyword
When things become complex, it will be a nightmare to fix the method
ambiguity with too many positional argument. Because keyword parameters
are not used in method dispatch, it is more friendly to both users and
maintainers.

This commit also fixes the incremental complilation issue; there are two
duplicated methods.